Solves the challenge, no bonuses yet.
Added file of allowed words, which are read into a std::unordered_set.
Now generate_permutations() only stores results to the results vector if
they are actual english words (meaning they are in the file of allowed
words).  The word file (enable1.txt) was provided as part of the
challenge.

While this program arrives at the correct answers, it is kind of slow
(~3 sec per set of tiles).  Solving the bonuses will require a more
efficient algorithm.